+++ Projects/ft_printf/printf/length_hh.c
@@ -31,15 +31,15 @@ char	*pull_un_char(t_print *form, va_list ap, char c)
 
 	arg = va_arg(ap, unsigned int);
 	if (c == 'u')
-		result = ft_unsigned_itoa((unsigned int)arg);
+		result = ft_unsigned_itoa(arg);
 	if (c == 'o')
-		result = ft_octal_convert((unsigned int)arg);
+		result = ft_octal_convert(arg);
 	if (c == 'x')
-		result = ft_hex_convert((unsigned int)arg);
+		result = ft_hex_convert(arg);
 	if (c == 'X')
 	{
 		i = 0;
-		result = ft_hex_convert((unsigned int)arg);
+		result = ft_hex_convert(arg);
 		while (result[i])
 		{
 			result[i] = (char)ft_toupper(result[i]);
